+++ MCBO/mcbo/acq_optimizers/interleaved_search_acq_optimizer.py
@@ -47,7 +47,7 @@ class InterleavedSearchAcqOptimizer(AcqOptimizerBase):
                  num_optimizer: str = 'adam',
                  num_lr: Optional[float] = None,
                  nominal_tol: int = 100,
-                 dtype: torch.dtype = torch.float32
+                 dtype: torch.dtype = torch.float64
                  ):
         super(InterleavedSearchAcqOptimizer, self).__init__(
             search_space=search_space,
